Adds feature to be able to insert employees to database

+++ system/src/main/java/com/navidasaman/system/repository/EmployeeRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository; // To implement data, access l
 import com.navidasaman.system.model.Employee; // To import employee from model 
 
 @Repository
-public interface EmployeeRepository extends JpaRepository<Employee, Integer> { // Employee from .model and primary key type which is an integer (id)
+public interface EmployeeRepository extends JpaRepository<Employee, Long> { // Employee from .model and primary key type which is an integer (id)
 
 }
